Eliminate “helper” class
